fixing bugs//connections btw modules
